requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
Pillow>=9.5.0
streamlit-extras>=0.3.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0